test: split config migration coverage by file
Keep the config migration regression coverage below the repository's per-test-file growth threshold without changing the exercised behavior.

Generated-By: looper 0.0.0-dev (runner=fixer, agent=opencode)
