+++ internal/cliapp/app_test.go
@@ -13,7 +13,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"reflect"
-	"slices"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -95,415 +94,6 @@ func TestCommandGroupHelpListsExpectedSubcommands(t *testing.T) {
 				}
 			}
 		})
-	}
-}
-
-func TestConfigMigrateDryRunPreviewsCanonicalTOMLWithoutWritingDestination(t *testing.T) {
-	homeDir := t.TempDir()
-	t.Setenv("HOME", homeDir)
-	fromPath := filepath.Join(homeDir, ".looper", "config.json")
-	toPath := filepath.Join(homeDir, ".looper", "config.toml")
-	if err := os.MkdirAll(filepath.Dir(fromPath), 0o755); err != nil {
-		t.Fatalf("MkdirAll() error = %v", err)
-	}
-	legacy := `{
-		"defaults": {"allowAutoApprove": true, "fixAllPullRequests": true},
-		"reviewer": {"reviewEvents": {"blocking": "REQUEST_CHANGES"}},
-		"roles": {"reviewer": {"autoDiscovery": true, "triggers": {"requireReviewRequest": true}, "specReview": {"includeReviewingLabel": true}}},
-		"projects": [{"id": "project_1", "name": "Repo", "path": "/tmp/repo", "instructions": {"reviewer": "be careful"}, "roles": {"reviewer": {"autoDiscovery": true, "triggers": {"requireReviewRequest": true}}}}],
-		"notifications": {"osascript": {"enabled": false}}
-	}`
-	if err := os.WriteFile(fromPath, []byte(legacy), 0o644); err != nil {
-		t.Fatalf("WriteFile() error = %v", err)
-	}
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--dry-run", "--from", fromPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --dry-run]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --dry-run]) stderr = %q, want empty", stderr)
-	}
-	if _, err := os.Stat(toPath); !os.IsNotExist(err) {
-		t.Fatalf("Stat(%s) err = %v, want not exists", toPath, err)
-	}
-	for _, notWant := range []string{"reviewer =", "[roles.reviewer.specReview]", "path = \"/tmp/repo\"", "instructions = {"} {
-		if strings.Contains(stdout, notWant) {
-			t.Fatalf("dry-run preview unexpectedly contained %q:\n%s", notWant, stdout)
-		}
-	}
-	for _, want := range []string{"Preview migration:", "[roles.reviewer.behavior.reviewEvents]", "clean = 'APPROVE'", "authorFilter = 'any'", "repoPath = '/tmp/repo'", "[projects.roles.reviewer.discovery]"} {
-		if !strings.Contains(stdout, want) {
-			t.Fatalf("dry-run preview missing %q:\n%s", want, stdout)
-		}
-	}
-
-	previewPath := filepath.Join(t.TempDir(), "preview.toml")
-	preview := stdout[strings.Index(stdout, "[roles.reviewer.behavior.reviewEvents]"):]
-	if err := os.WriteFile(previewPath, []byte(preview), 0o644); err != nil {
-		t.Fatalf("WriteFile(preview) error = %v", err)
-	}
-	loaded, err := config.LoadFile(config.LoadFileOptions{CWD: t.TempDir(), ConfigPath: previewPath, LookupEnv: emptyConfigEnvLookup, LookPath: configLookPathForTests()})
-	if err != nil {
-		t.Fatalf("LoadFile(preview) error = %v", err)
-	}
-	if len(loaded.Warnings) != 0 {
-		t.Fatalf("LoadFile(preview).Warnings = %#v, want none", loaded.Warnings)
-	}
-	if len(loaded.Notices) != 0 {
-		t.Fatalf("LoadFile(preview).Notices = %#v, want none", loaded.Notices)
-	}
-}
-
-func TestConfigMigrateWritesDestinationAndPreservesSource(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	toPath := filepath.Join(filepath.Dir(fromPath), "config.toml")
-	beforeSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source) error = %v", err)
-	}
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--from", fromPath, "--to", toPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate]) stderr = %q, want empty", stderr)
-	}
-	if !strings.Contains(stdout, "Migrated config:") {
-		t.Fatalf("stdout = %q, want migrated summary", stdout)
-	}
-	afterSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source after) error = %v", err)
-	}
-	if !bytes.Equal(beforeSource, afterSource) {
-		t.Fatal("source config changed during migration")
-	}
-	rawDest, err := os.ReadFile(toPath)
-	if err != nil {
-		t.Fatalf("ReadFile(dest) error = %v", err)
-	}
-	if strings.Contains(string(rawDest), "allowAutoApprove") {
-		t.Fatalf("destination still contained legacy key:\n%s", rawDest)
-	}
-}
-
-func TestConfigMigrateDryRunJSONReportsLegacyChanges(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{
-		"defaults": map[string]any{"allowAutoApprove": true},
-		"projects": []map[string]any{{
-			"id":   "project_1",
-			"name": "Repo",
-			"path": "/tmp/repo",
-			"instructions": map[string]any{
-				"worker": "legacy worker instructions",
-			},
-		}},
-		"notifications": map[string]any{"osascript": map[string]any{"enabled": false}},
-	})
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--dry-run", "--json", "--from", fromPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --dry-run --json]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --dry-run --json]) stderr = %q, want empty", stderr)
-	}
-
-	var decoded map[string]any
-	if err := json.Unmarshal([]byte(stdout), &decoded); err != nil {
-		t.Fatalf("Unmarshal(stdout) error = %v\nraw=%q", err, stdout)
-	}
-	changes, ok := decoded["changes"].([]any)
-	if !ok {
-		t.Fatalf("changes type = %T, want []any", decoded["changes"])
-	}
-	got := make([]string, 0, len(changes))
-	for _, change := range changes {
-		text, ok := change.(string)
-		if !ok {
-			t.Fatalf("change type = %T, want string", change)
-		}
-		got = append(got, text)
-	}
-	for _, want := range []string{
-		"defaults.allowAutoApprove -> roles.reviewer.behavior.reviewEvents.clean",
-		"projects[0].instructions -> projects[0].roles.<role>.instructions",
-	} {
-		if !slices.Contains(got, want) {
-			t.Fatalf("changes = %#v, want %q", got, want)
-		}
-	}
-}
-
-func TestConfigMigrateExistingDestinationRequiresForce(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	toPath := filepath.Join(filepath.Dir(fromPath), "config.toml")
-	if err := os.WriteFile(toPath, []byte("existing = true\n"), 0o644); err != nil {
-		t.Fatalf("WriteFile(dest) error = %v", err)
-	}
-
-	exitCode, _, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--from", fromPath, "--to", toPath)
-	if exitCode == 0 {
-		t.Fatalf("Run([config migrate]) exit code = 0, want non-zero")
-	}
-	if !strings.Contains(stderr, "destination already exists") {
-		t.Fatalf("stderr = %q, want destination exists error", stderr)
-	}
-}
-
-func TestConfigMigrateForceCreatesBackup(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	toPath := filepath.Join(filepath.Dir(fromPath), "config.toml")
-	if err := os.WriteFile(toPath, []byte("existing = true\n"), 0o644); err != nil {
-		t.Fatalf("WriteFile(dest) error = %v", err)
-	}
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--force", "--json", "--from", fromPath, "--to", toPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --force --json]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --force --json]) stderr = %q, want empty", stderr)
-	}
-	var decoded map[string]any
-	if err := json.Unmarshal([]byte(stdout), &decoded); err != nil {
-		t.Fatalf("Unmarshal(stdout) error = %v\nraw=%q", err, stdout)
-	}
-	backupPath, _ := decoded["backupPath"].(string)
-	if backupPath == "" {
-		t.Fatalf("backupPath missing from JSON output: %#v", decoded)
-	}
-	backupRaw, err := os.ReadFile(backupPath)
-	if err != nil {
-		t.Fatalf("ReadFile(backup) error = %v", err)
-	}
-	if string(backupRaw) != "existing = true\n" {
-		t.Fatalf("backup contents = %q, want original destination", backupRaw)
-	}
-}
-
-func TestConfigMigrateDanglingSymlinkDestinationRequiresForce(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	toPath := filepath.Join(filepath.Dir(fromPath), "config.toml")
-	if err := os.Symlink(filepath.Join(t.TempDir(), "missing.toml"), toPath); err != nil {
-		t.Fatalf("Symlink(dest) error = %v", err)
-	}
-
-	exitCode, _, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--from", fromPath, "--to", toPath)
-	if exitCode == 0 {
-		t.Fatalf("Run([config migrate]) exit code = 0, want non-zero")
-	}
-	if !strings.Contains(stderr, "destination already exists") {
-		t.Fatalf("stderr = %q, want destination exists error", stderr)
-	}
-}
-
-func TestConfigMigrateForceBacksUpDanglingSymlinkDestination(t *testing.T) {
-	t.Parallel()
-
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	toPath := filepath.Join(filepath.Dir(fromPath), "config.toml")
-	linkTarget := filepath.Join(t.TempDir(), "missing.toml")
-	if err := os.Symlink(linkTarget, toPath); err != nil {
-		t.Fatalf("Symlink(dest) error = %v", err)
-	}
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--force", "--json", "--from", fromPath, "--to", toPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --force --json]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --force --json]) stderr = %q, want empty", stderr)
-	}
-	var decoded map[string]any
-	if err := json.Unmarshal([]byte(stdout), &decoded); err != nil {
-		t.Fatalf("Unmarshal(stdout) error = %v\nraw=%q", err, stdout)
-	}
-	backupPath, _ := decoded["backupPath"].(string)
-	if backupPath == "" {
-		t.Fatalf("backupPath missing from JSON output: %#v", decoded)
-	}
-	backupTarget, err := os.Readlink(backupPath)
-	if err != nil {
-		t.Fatalf("Readlink(backup) error = %v", err)
-	}
-	if backupTarget != linkTarget {
-		t.Fatalf("backup symlink target = %q, want %q", backupTarget, linkTarget)
-	}
-	if info, err := os.Lstat(toPath); err != nil {
-		t.Fatalf("Lstat(dest after) error = %v", err)
-	} else if info.Mode()&os.ModeSymlink != 0 {
-		t.Fatalf("destination remained symlink after migration")
-	}
-}
-
-func TestBackupConfigFileCopiesSymlinkTargetContents(t *testing.T) {
-	t.Parallel()
-
-	dir := t.TempDir()
-	targetPath := filepath.Join(dir, "real-config.toml")
-	if err := os.WriteFile(targetPath, []byte("before = true\n"), 0o644); err != nil {
-		t.Fatalf("WriteFile(target) error = %v", err)
-	}
-	configPath := filepath.Join(dir, "config.toml")
-	if err := os.Symlink(targetPath, configPath); err != nil {
-		t.Fatalf("Symlink(config) error = %v", err)
-	}
-
-	if err := backupConfigFile(configPath); err != nil {
-		t.Fatalf("backupConfigFile() error = %v", err)
-	}
-
-	backupPaths, err := filepath.Glob(configPath + ".*.bak")
-	if err != nil {
-		t.Fatalf("Glob(backup) error = %v", err)
-	}
-	if len(backupPaths) != 1 {
-		t.Fatalf("len(backupPaths) = %d, want 1", len(backupPaths))
-	}
-	if info, err := os.Lstat(backupPaths[0]); err != nil {
-		t.Fatalf("Lstat(backup) error = %v", err)
-	} else if info.Mode()&os.ModeSymlink != 0 {
-		t.Fatalf("backup remained symlink")
-	}
-	if err := os.WriteFile(targetPath, []byte("after = true\n"), 0o644); err != nil {
-		t.Fatalf("WriteFile(target after) error = %v", err)
-	}
-	backupRaw, err := os.ReadFile(backupPaths[0])
-	if err != nil {
-		t.Fatalf("ReadFile(backup) error = %v", err)
-	}
-	if string(backupRaw) != "before = true\n" {
-		t.Fatalf("backup contents = %q, want original target contents", backupRaw)
-	}
-}
-
-func TestConfigMigrateUsesActiveConfigPathFromFlagWhenFromOmitted(t *testing.T) {
-	homeDir := t.TempDir()
-	t.Setenv("HOME", homeDir)
-	envPath := writeEditableCLIConfigWithPayload(t, map[string]any{"defaults": map[string]any{"allowAutoApprove": false}, "notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	t.Setenv("LOOPER_CONFIG", envPath)
-
-	rootDir := t.TempDir()
-	fromPath := filepath.Join(rootDir, "custom", "active.json")
-	if err := os.MkdirAll(filepath.Dir(fromPath), 0o755); err != nil {
-		t.Fatalf("MkdirAll() error = %v", err)
-	}
-	if err := os.WriteFile(fromPath, []byte(`{"notifications":{"osascript":{"enabled":false}}}`), 0o644); err != nil {
-		t.Fatalf("WriteFile() error = %v", err)
-	}
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--dry-run", "--config", fromPath)
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --dry-run --config]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --dry-run --config]) stderr = %q, want empty", stderr)
-	}
-	want := "Preview migration: " + fromPath + " -> " + strings.TrimSuffix(fromPath, filepath.Ext(fromPath)) + ".toml"
-	if !strings.Contains(stdout, want) {
-		t.Fatalf("stdout = %q, want %q", stdout, want)
-	}
-}
-
-func TestConfigMigrateUsesActiveConfigPathFromEnvWhenFromOmitted(t *testing.T) {
-	homeDir := t.TempDir()
-	t.Setenv("HOME", homeDir)
-	fromPath := filepath.Join(t.TempDir(), "env", "active.json")
-	if err := os.MkdirAll(filepath.Dir(fromPath), 0o755); err != nil {
-		t.Fatalf("MkdirAll() error = %v", err)
-	}
-	if err := os.WriteFile(fromPath, []byte(`{"notifications":{"osascript":{"enabled":false}}}`), 0o644); err != nil {
-		t.Fatalf("WriteFile() error = %v", err)
-	}
-	t.Setenv("LOOPER_CONFIG", fromPath)
-
-	exitCode, stdout, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--dry-run")
-	if exitCode != 0 {
-		t.Fatalf("Run([config migrate --dry-run]) exit code = %d; stdout=%q stderr=%q", exitCode, stdout, stderr)
-	}
-	if stderr != "" {
-		t.Fatalf("Run([config migrate --dry-run]) stderr = %q, want empty", stderr)
-	}
-	want := "Preview migration: " + fromPath + " -> " + strings.TrimSuffix(fromPath, filepath.Ext(fromPath)) + ".toml"
-	if !strings.Contains(stdout, want) {
-		t.Fatalf("stdout = %q, want %q", stdout, want)
-	}
-}
-
-func TestConfigMigrateRejectsSameFileAcrossAbsoluteAndRelativePaths(t *testing.T) {
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	fromDir := filepath.Dir(fromPath)
-	fromBase := filepath.Base(fromPath)
-	beforeSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source) error = %v", err)
-	}
-	wd, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Getwd() error = %v", err)
-	}
-	defer func() {
-		if err := os.Chdir(wd); err != nil {
-			t.Fatalf("Chdir(restore) error = %v", err)
-		}
-	}()
-	if err := os.Chdir(fromDir); err != nil {
-		t.Fatalf("Chdir(%s) error = %v", fromDir, err)
-	}
-
-	exitCode, _, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--force", "--from", fromPath, "--to", fromBase)
-	if exitCode == 0 {
-		t.Fatalf("Run([config migrate --force]) exit code = 0, want non-zero")
-	}
-	if !strings.Contains(stderr, "source and destination must differ") {
-		t.Fatalf("stderr = %q, want same-file error", stderr)
-	}
-	afterSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source after) error = %v", err)
-	}
-	if !bytes.Equal(beforeSource, afterSource) {
-		t.Fatal("source config changed during rejected migration")
-	}
-}
-
-func TestConfigMigrateRejectsSameFileThroughSymlinkAlias(t *testing.T) {
-	fromPath := writeEditableCLIConfigWithPayload(t, map[string]any{"notifications": map[string]any{"osascript": map[string]any{"enabled": false}}})
-	aliasPath := filepath.Join(filepath.Dir(fromPath), "config-alias.json")
-	if err := os.Symlink(fromPath, aliasPath); err != nil {
-		t.Fatalf("Symlink(%s, %s) error = %v", fromPath, aliasPath, err)
-	}
-	beforeSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source) error = %v", err)
-	}
-
-	exitCode, _, stderr := runAppWithLookPath(t, configLookPathForTests(), "config", "migrate", "--force", "--from", fromPath, "--to", aliasPath)
-	if exitCode == 0 {
-		t.Fatalf("Run([config migrate --force]) exit code = 0, want non-zero")
-	}
-	if !strings.Contains(stderr, "source and destination must differ") {
-		t.Fatalf("stderr = %q, want same-file error", stderr)
-	}
-	afterSource, err := os.ReadFile(fromPath)
-	if err != nil {
-		t.Fatalf("ReadFile(source after) error = %v", err)
-	}
-	if !bytes.Equal(beforeSource, afterSource) {
-		t.Fatal("source config changed during rejected migration")
 	}
 }
 
@@ -4264,16 +3854,6 @@ func writeEnvelope(t *testing.T, w http.ResponseWriter, payload any) {
 	if err := json.NewEncoder(w).Encode(payload); err != nil {
 		t.Fatalf("encode envelope: %v", err)
 	}
-}
-
-func configLookPathForTests() config.LookPathFunc {
-	return func(name string) (string, error) {
-		return filepath.Join(string(os.PathSeparator), "detected", name), nil
-	}
-}
-
-func emptyConfigEnvLookup(string) (string, bool) {
-	return "", false
 }
 
 func assertJSONContains(t *testing.T, raw string, key string, want any) {
